+++ packages/jsdoc-plugins/CHANGELOG.md
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+## [1.3.0](https://github.com/ckeditor/ckeditor5-dev/compare/@ckeditor/jsdoc-plugins@1.2.1...@ckeditor/jsdoc-plugins@1.3.0) (2017-05-18)
+
+### Features
+
+* Introduced JSDoc plugin enabling inheritance of static members. Closes [#201](https://github.com/ckeditor/ckeditor5-dev/issues/201). ([08c0a21](https://github.com/ckeditor/ckeditor5-dev/commit/08c0a21))
+
+
 ## [1.2.1](https://github.com/ckeditor/ckeditor5-dev/compare/@ckeditor/jsdoc-plugins@1.2.0...@ckeditor/jsdoc-plugins@1.2.1) (2017-04-27)
 
 Internal changes only (updated dependencies, documentation, etc.).